Move fetch hash into library
